+++ integration_tests/generics_array_01.py
@@ -1,5 +1,5 @@
 from lpython import TypeVar, i32
-from numpy import empty
+from numpy import empty, int32
 
 T = TypeVar('T')
 
@@ -9,7 +9,7 @@ def f(lst: T[:], i: T) -> T:
 
 def use_array():
     array: i32[1]
-    array = empty(1)
+    array = empty(1, dtype=int32)
     x: i32
     x = 69
     print(f(array, x))
